+++ Wire-iOS/Sources/UserInterface/Registration/Authentication/Coordinator/AuthenticationCoordinator+Navigation.swift
@@ -1,0 +1,44 @@
+//
+// Wire
+// Copyright (C) 2018 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import UIKit
+
+extension AuthenticationCoordinator: UINavigationControllerDelegate {
+
+    func navigationController(_ navigationController: UINavigationController, didShow viewController: UIViewController, animated: Bool) {
+        guard let currentViewController = self.currentViewController else {
+            return
+        }
+
+        guard let keyboardViewController = viewController as? KeyboardAvoidingViewController else {
+            return
+        }
+
+        guard let authenticationViewController = keyboardViewController.viewController as? AuthenticationStepViewController else {
+            return
+        }
+
+        guard authenticationViewController.isEqual(currentViewController) == false else {
+            return
+        }
+
+        self.currentViewController = authenticationViewController
+        unwind(requireInterfaceStep: true)
+    }
+
+}
